Fixed windows compilation errors and untabified a few files

+++ Engine/Entities/GameWorld.cpp
@@ -62,27 +62,27 @@ DLLEXPORT bool Leviathan::GameWorld::Init(NETWORKED_TYPE type, GraphicalInputEnt
 {
     IsOnServer = (type == NETWORKED_TYPE::Server);
     
-	LinkedToWindow = renderto;
+    LinkedToWindow = renderto;
 
-	// Detecting non-GUI mode //
-	if(ogre){
+    // Detecting non-GUI mode //
+    if(ogre){
         
-		if(!renderto)
-			return false;
+        if(!renderto)
+            return false;
         
-		GraphicalMode = true;
-		// these are always required for worlds //
-		_CreateOgreResources(ogre, renderto);
-	}
+        GraphicalMode = true;
+        // these are always required for worlds //
+        _CreateOgreResources(ogre, renderto);
+    }
 
-	// Acquire physics engine world //
+    // Acquire physics engine world //
     // This should not be required if it isn't available
     if(NewtonManager::Get()){
         
         _PhysicalWorld = NewtonManager::Get()->CreateWorld(this);
     }
 
-	return true;
+    return true;
 }
 
 DLLEXPORT void Leviathan::GameWorld::Release(){
@@ -91,83 +91,83 @@ DLLEXPORT void Leviathan::GameWorld::Release(){
     
     ReceivingPlayers.clear();
     
-	// release objects //
+    // release objects //
     // TODO: allow objects to know that they are about to get killed
 
     // As all objects are just pointers to components we can just dump the objects
     // and once the component pools are released
     ClearObjects();
     
-	if(GraphicalMode){
-		// TODO: notify our window that it no longer has a world workspace
-		LinkedToWindow = NULL;
+    if(GraphicalMode){
+        // TODO: notify our window that it no longer has a world workspace
+        LinkedToWindow = NULL;
 
-		// release Ogre resources //
+        // release Ogre resources //
 
-		// Destroy the compositor //
-		Ogre::Root& ogre = Ogre::Root::getSingleton();
-		ogre.getCompositorManager2()->removeWorkspace(WorldWorkspace);
-		WorldWorkspace = NULL;
-		ogre.destroySceneManager(WorldsScene);
-		WorldsScene = NULL;
-	}
+        // Destroy the compositor //
+        Ogre::Root& ogre = Ogre::Root::getSingleton();
+        ogre.getCompositorManager2()->removeWorkspace(WorldWorkspace);
+        WorldWorkspace = NULL;
+        ogre.destroySceneManager(WorldsScene);
+        WorldsScene = NULL;
+    }
     
 }
 // ------------------------------------ //
 void Leviathan::GameWorld::_CreateOgreResources(Ogre::Root* ogre,
     GraphicalInputEntity* rendertarget)
 {
-	// create scene manager //
-	WorldsScene = ogre->createSceneManager(Ogre::ST_EXTERIOR_FAR, 2,
+    // create scene manager //
+    WorldsScene = ogre->createSceneManager(Ogre::ST_EXTERIOR_FAR, 2,
         Ogre::INSTANCING_CULLING_THREADED,
         "MainSceneManager_"+Convert::ToString(ID));
 
-	// WorldsScene->setShadowFarDistance(1000.f);
-	// WorldsScene->setShadowDirectionalLightExtrusionDistance(10000.f);
+    // WorldsScene->setShadowFarDistance(1000.f);
+    // WorldsScene->setShadowDirectionalLightExtrusionDistance(10000.f);
 
-	// create camera //
-	WorldSceneCamera = WorldsScene->createCamera("Camera01");
+    // create camera //
+    WorldSceneCamera = WorldsScene->createCamera("Camera01");
 
     // WorldSceneCamera->lookAt( Ogre::Vector3( 0, 0, 0 ) );
     
-	// near and far clipping planes //
-	WorldSceneCamera->setFOVy(Ogre::Degree(60));
-	WorldSceneCamera->setNearClipDistance(0.1f);
-	WorldSceneCamera->setFarClipDistance(50000.f);
+    // near and far clipping planes //
+    WorldSceneCamera->setFOVy(Ogre::Degree(60));
+    WorldSceneCamera->setNearClipDistance(0.1f);
+    WorldSceneCamera->setFarClipDistance(50000.f);
 
-	// enable infinite far clip distance if supported //
-	if(ogre->getRenderSystem()->getCapabilities()->hasCapability(
+    // enable infinite far clip distance if supported //
+    if(ogre->getRenderSystem()->getCapabilities()->hasCapability(
             Ogre::RSC_INFINITE_FAR_PLANE))
     {
-		
-		WorldSceneCamera->setFarClipDistance(0);
-	}
+        
+        WorldSceneCamera->setFarClipDistance(0);
+    }
 
-	// set scene ambient colour //
+    // set scene ambient colour //
     // TODO: update this to the PBR HLMS system
-	// WorldsScene->setAmbientLight(Ogre::ColourValue(0.3f, 0.3f, 0.3f),
+    // WorldsScene->setAmbientLight(Ogre::ColourValue(0.3f, 0.3f, 0.3f),
     //     Ogre::ColourValue(0.2f, 0.2f, 0.2f), Ogre::Vector3(0.1f, 1.f, 0.085f));
 
-	// // default sun //
-	SetSunlight();
-	
-	// Create the workspace for this scene //
-	// Which will be rendered before the overlay workspace but after potential
+    // // default sun //
+    SetSunlight();
+    
+    // Create the workspace for this scene //
+    // Which will be rendered before the overlay workspace but after potential
     // clearing workspace
-	WorldWorkspace = ogre->getCompositorManager2()->addWorkspace(WorldsScene,
+    WorldWorkspace = ogre->getCompositorManager2()->addWorkspace(WorldsScene,
         rendertarget->GetOgreWindow(), 
-		WorldSceneCamera, "WorldsWorkspace", true, 0);
+        WorldSceneCamera, "WorldsWorkspace", true, 0);
 
 }
 // ------------------------------------ //
 DLLEXPORT void Leviathan::GameWorld::SetSkyBox(const std::string &materialname){
-	try{
-		WorldsScene->setSkyBox(true, materialname);
-	}
-	catch(const Ogre::InvalidParametersException &e){
+    try{
+        WorldsScene->setSkyBox(true, materialname);
+    }
+    catch(const Ogre::InvalidParametersException &e){
 
-		Logger::Get()->Error("[EXCEPTION] "+e.getFullDescription());
-	}
+        Logger::Get()->Error("[EXCEPTION] "+e.getFullDescription());
+    }
 }
 // ------------------------------------ //
 DLLEXPORT void GameWorld::SetSkyPlane(const std::string &material, const Ogre::Plane &plane
@@ -182,41 +182,41 @@ DLLEXPORT void GameWorld::DisableSkyPlane(){
 }
 // ------------------------------------ //
 DLLEXPORT void Leviathan::GameWorld::SetFog(){
-	WorldsScene->setFog(Ogre::FOG_LINEAR, Ogre::ColourValue(0.7f, 0.7f, 0.8f), 0, 4000, 10000);
-	WorldsScene->setFog(Ogre::FOG_NONE);
+    WorldsScene->setFog(Ogre::FOG_LINEAR, Ogre::ColourValue(0.7f, 0.7f, 0.8f), 0, 4000, 10000);
+    WorldsScene->setFog(Ogre::FOG_NONE);
 }
 
 DLLEXPORT void Leviathan::GameWorld::SetSunlight(){
-	// create/update things if they are NULL //
-	if(!Sunlight){
-		Sunlight = WorldsScene->createLight();
-		Sunlight->setName("sunlight");
-	}
+    // create/update things if they are NULL //
+    if(!Sunlight){
+        Sunlight = WorldsScene->createLight();
+        Sunlight->setName("sunlight");
+    }
 
-	Sunlight->setType(Ogre::Light::LT_DIRECTIONAL);
-	Sunlight->setDiffuseColour(0.98f, 1.f, 0.95f);
-	Sunlight->setSpecularColour(1.f, 1.f, 1.f);
+    Sunlight->setType(Ogre::Light::LT_DIRECTIONAL);
+    Sunlight->setDiffuseColour(0.98f, 1.f, 0.95f);
+    Sunlight->setSpecularColour(1.f, 1.f, 1.f);
 
-	if(!SunLightNode){
+    if(!SunLightNode){
 
-		SunLightNode = WorldsScene->getRootSceneNode()->createChildSceneNode();
-		SunLightNode->setName("sunlight node");
+        SunLightNode = WorldsScene->getRootSceneNode()->createChildSceneNode();
+        SunLightNode->setName("sunlight node");
 
-		SunLightNode->attachObject(Sunlight);
-	}
+        SunLightNode->attachObject(Sunlight);
+    }
 
-	Ogre::Quaternion quat;
-	quat.FromAngleAxis(Ogre::Radian(1.f), Float3(0.55f, -0.3f, 0.75f));
-	SunLightNode->setOrientation(quat);
+    Ogre::Quaternion quat;
+    quat.FromAngleAxis(Ogre::Radian(1.f), Float3(0.55f, -0.3f, 0.75f));
+    SunLightNode->setOrientation(quat);
 }
 
 DLLEXPORT void Leviathan::GameWorld::RemoveSunlight(){
-	if(SunLightNode){
-		SunLightNode->detachAllObjects();
-		// might be safe to delete
-		OGRE_DELETE SunLightNode;
-		SunLightNode = NULL;
-	}
+    if(SunLightNode){
+        SunLightNode->detachAllObjects();
+        // might be safe to delete
+        OGRE_DELETE SunLightNode;
+        SunLightNode = NULL;
+    }
 }
 
 DLLEXPORT void Leviathan::GameWorld::Render(int mspassed, int tick, int timeintick){
@@ -536,10 +536,10 @@ DLLEXPORT float Leviathan::GameWorld::GetTickProgress() const {
     return progress < 1.f ? progress : 1.f;
 }
 
-DLLEXPORT std::tuple<int, int64_t> GameWorld::GetTickAndTime() const{
+DLLEXPORT std::tuple<int, int> GameWorld::GetTickAndTime() const{
 
     int tick = TickNumber;
-    int timeSince = Engine::Get()->GetTimeSinceLastTick();
+    int timeSince = static_cast<int>(Engine::Get()->GetTimeSinceLastTick());
 
     while(timeSince >= TICKSPEED){
 
@@ -636,9 +636,9 @@ DLLEXPORT void Leviathan::GameWorld::ClearObjects(){
 }
 // ------------------------------------ //
 DLLEXPORT Float3 Leviathan::GameWorld::GetGravityAtPosition(const Float3 &pos){
-	// \todo take position into account //
-	// create force without mass applied //
-	return Float3(0.f, PHYSICS_BASE_GRAVITY, 0.f);
+    // \todo take position into account //
+    // create force without mass applied //
+    return Float3(0.f, PHYSICS_BASE_GRAVITY, 0.f);
 }
 // ------------------------------------ //
 DLLEXPORT int Leviathan::GameWorld::GetPhysicalMaterial(const std::string &name){
@@ -650,79 +650,79 @@ DLLEXPORT int Leviathan::GameWorld::GetPhysicalMaterial(const std::string &name)
 DLLEXPORT void Leviathan::GameWorld::DestroyObject(ObjectID id){
 
     auto end = Objects.end();
-	for(auto iter = Objects.begin(); iter != end; ++iter){
+    for(auto iter = Objects.begin(); iter != end; ++iter){
         
-		if(*iter == id){
+        if(*iter == id){
 
             _DoDestroy(id);
-			Objects.erase(iter);
+            Objects.erase(iter);
             
-			return;
-		}
-	}
+            return;
+        }
+    }
 }
 
 DLLEXPORT void Leviathan::GameWorld::QueueDestroyObject(ObjectID id) {
 
-	Lock lock(DeleteMutex);
-	DelayedDeleteIDS.push_back(id);
+    Lock lock(DeleteMutex);
+    DelayedDeleteIDS.push_back(id);
 }
 
 void Leviathan::GameWorld::_HandleDelayedDelete(){
 
-	// We might want to delete everything //
-	if(ClearAllObjects){
+    // We might want to delete everything //
+    if(ClearAllObjects){
 
-		ClearObjects();
+        ClearObjects();
 
-		ClearAllObjects = false;
+        ClearAllObjects = false;
 
         Lock lock(DeleteMutex);
         DelayedDeleteIDS.clear();
         
-		// All are now cleared //
-		return;
-	}
+        // All are now cleared //
+        return;
+    }
 
     Lock lock(DeleteMutex);
     
-	// Return right away if no objects to delete //
-	if(DelayedDeleteIDS.empty())
-		return;
+    // Return right away if no objects to delete //
+    if(DelayedDeleteIDS.empty())
+        return;
 
-	// Search all objects and find the ones that need to be deleted //
-	for(auto iter = Objects.begin(); iter != Objects.end(); ){
+    // Search all objects and find the ones that need to be deleted //
+    for(auto iter = Objects.begin(); iter != Objects.end(); ){
 
-		// Check does id match any //
-		auto curid = *iter;
-		bool delthis = false;
+        // Check does id match any //
+        auto curid = *iter;
+        bool delthis = false;
 
-		for(auto iterids = DelayedDeleteIDS.begin(); iterids != DelayedDeleteIDS.end(); ){
+        for(auto iterids = DelayedDeleteIDS.begin(); iterids != DelayedDeleteIDS.end(); ){
 
-			if(*iterids == curid){
-				// Remove this as it will get deleted //
-				delthis = true;
-				DelayedDeleteIDS.erase(iterids);
-				break;
+            if(*iterids == curid){
+                // Remove this as it will get deleted //
+                delthis = true;
+                DelayedDeleteIDS.erase(iterids);
+                break;
 
-			} else {
-				++iterids;
-			}
-		}
+            } else {
+                ++iterids;
+            }
+        }
 
-		if(delthis){
+        if(delthis){
 
             _DoDestroy(curid);
-			iter = Objects.erase(iter);
+            iter = Objects.erase(iter);
             
-			// Check for end //
-			if(DelayedDeleteIDS.empty())
+            // Check for end //
+            if(DelayedDeleteIDS.empty())
                 return;
 
-		} else {
-			++iter;
-		}
-	}
+        } else {
+            ++iter;
+        }
+    }
 }
 
 void GameWorld::_DoDestroy(ObjectID id){
@@ -763,11 +763,11 @@ void Leviathan::GameWorld::_ReportEntityDestruction(ObjectID id){
 
 // ------------------------------------ //
 DLLEXPORT void Leviathan::GameWorld::SetWorldPhysicsFrozenState(bool frozen){
-	// Skip if set to the same //
-	if(frozen == WorldFrozen)
-		return;
+    // Skip if set to the same //
+    if(frozen == WorldFrozen)
+        return;
 
-	WorldFrozen = frozen;
+    WorldFrozen = frozen;
 
     // Send it to receiving players (if we are a server) //
     if(ReceivingPlayers.empty())
@@ -781,25 +781,25 @@ DLLEXPORT void Leviathan::GameWorld::SetWorldPhysicsFrozenState(bool frozen){
 DLLEXPORT RayCastHitEntity* Leviathan::GameWorld::CastRayGetFirstHit(const Float3 &from,
     const Float3 &to)
 {
-	// Create a data object for the ray cast //
-	RayCastData data(1, from, to);
+    // Create a data object for the ray cast //
+    RayCastData data(1, from, to);
 
-	// Call the actual ray firing function //
-	NewtonWorldRayCast(_PhysicalWorld->GetNewtonWorld(), &from.X, &to.X,
+    // Call the actual ray firing function //
+    NewtonWorldRayCast(_PhysicalWorld->GetNewtonWorld(), &from.X, &to.X,
         RayCallbackDataCallbackClosest, &data, NULL, 0);
 
-	// Check the result //
-	if(data.HitEntities.size() == 0){
-		// Nothing hit //
-		return new RayCastHitEntity();
-	}
+    // Check the result //
+    if(data.HitEntities.size() == 0){
+        // Nothing hit //
+        return new RayCastHitEntity();
+    }
     
-	// We need to increase reference count to not to accidentally delete the result while
+    // We need to increase reference count to not to accidentally delete the result while
     // caller is using it
-	data.HitEntities[0]->AddRef();
+    data.HitEntities[0]->AddRef();
     
-	// Return the only hit //
-	return data.HitEntities[0];
+    // Return the only hit //
+    return data.HitEntities[0];
 }
 
 // \todo improve this performance //
@@ -808,31 +808,31 @@ dFloat Leviathan::GameWorld::RayCallbackDataCallbackClosest(const NewtonBody* co
     const dFloat* const hitNormal, dLong collisionID, void* const userData,
     dFloat intersectParam)
 {
-	// Let's just store it as a NewtonBody pointer //
-	RayCastData* data = reinterpret_cast<RayCastData*>(userData);
+    // Let's just store it as a NewtonBody pointer //
+    RayCastData* data = reinterpret_cast<RayCastData*>(userData);
 
-	if(data->HitEntities.size() == 0)
-		data->HitEntities.push_back(new RayCastHitEntity(body, intersectParam, data));
-	else
-		*data->HitEntities[0] = RayCastHitEntity(body, intersectParam, data);
+    if(data->HitEntities.size() == 0)
+        data->HitEntities.push_back(new RayCastHitEntity(body, intersectParam, data));
+    else
+        *data->HitEntities[0] = RayCastHitEntity(body, intersectParam, data);
 
-	// Continue //
-	return intersectParam;
+    // Continue //
+    return intersectParam;
 }
 
 DLLEXPORT RayCastHitEntity* Leviathan::GameWorld::CastRayGetFirstHitProxy(const Float3 &from,
     const Float3 &to)
 {
-	return CastRayGetFirstHit(from, to);
+    return CastRayGetFirstHit(from, to);
 }
 
 DLLEXPORT void Leviathan::GameWorld::MarkForClear(){
-	ClearAllObjects = true;
+    ClearAllObjects = true;
 }
 // ------------------------------------ //
 void Leviathan::GameWorld::UpdatePlayersPositionData(ConnectedPlayer &ply)
 {
-	// Get the position for this player in this world //
+    // Get the position for this player in this world //
     ObjectID id = ply.GetPositionInWorld(this);
 
     // Player is using a static position at (0, 0, 0) //
@@ -1067,47 +1067,47 @@ DLLEXPORT void Leviathan::GameWorld::_OnCustomComponentDestroyed(ObjectID id,
 DLLEXPORT Leviathan::RayCastHitEntity::RayCastHitEntity(const NewtonBody* ptr /*= NULL*/,
     const float &tvar, RayCastData* ownerptr) : HitEntity(ptr), HitVariable(tvar)
 {
-	if(ownerptr){
-		HitLocation = ownerptr->BaseHitLocationCalcVar*HitVariable;
-	} else {
-		HitLocation = (Float3)0;
-	}
+    if(ownerptr){
+        HitLocation = ownerptr->BaseHitLocationCalcVar*HitVariable;
+    } else {
+        HitLocation = (Float3)0;
+    }
 }
 
 DLLEXPORT bool Leviathan::RayCastHitEntity::HasHit(){
-	return HitEntity != NULL;
+    return HitEntity != NULL;
 }
 
 DLLEXPORT bool Leviathan::RayCastHitEntity::DoesBodyMatchThisHit(NewtonBody* other){
-	return HitEntity == other;
+    return HitEntity == other;
 }
 
 DLLEXPORT Float3 Leviathan::RayCastHitEntity::GetPosition(){
-	return HitLocation;
+    return HitLocation;
 }
 
 DLLEXPORT RayCastHitEntity& Leviathan::RayCastHitEntity::operator=(
     const RayCastHitEntity& other)
 {
-	HitEntity = other.HitEntity;
-	HitVariable = other.HitVariable;
-	HitLocation = other.HitLocation;
+    HitEntity = other.HitEntity;
+    HitVariable = other.HitVariable;
+    HitLocation = other.HitLocation;
 
-	return *this;
+    return *this;
 }
 // ------------------ RayCastData ------------------ //
 DLLEXPORT Leviathan::RayCastData::RayCastData(int maxcount, const Float3 &from,
     const Float3 &to) :
     MaxCount(maxcount), 
-	// Formula based on helpful guy on Newton forums
-	BaseHitLocationCalcVar(from+(to-from))
+    // Formula based on helpful guy on Newton forums
+    BaseHitLocationCalcVar(from+(to-from))
 {
-	// Reserve memory for maximum number of objects //
-	HitEntities.reserve(maxcount);
+    // Reserve memory for maximum number of objects //
+    HitEntities.reserve(maxcount);
 }
 
 DLLEXPORT Leviathan::RayCastData::~RayCastData(){
-	// We want to release all hit data //
-	SAFE_RELEASE_VECTOR(HitEntities);
+    // We want to release all hit data //
+    SAFE_RELEASE_VECTOR(HitEntities);
 }
